+++ dflib/src/main/java/org/dflib/series/IntBaseSeries.java
@@ -297,7 +297,7 @@ public abstract class IntBaseSeries implements IntSeries {
     @Override
     public void copyTo(Object[] to, int fromOffset, int toOffset, int len) {
         for (int i = 0; i < len; i++) {
-            to[toOffset + i] = getInt(i);
+            to[toOffset + i] = getInt(fromOffset + i);
         }
     }